+++ src/main/java/com/gdschongik/gdsc/domain/membership/application/MembershipService.java
@@ -7,7 +7,6 @@ import com.gdschongik.gdsc.domain.membership.dao.MembershipRepository;
 import com.gdschongik.gdsc.domain.membership.domain.Membership;
 import com.gdschongik.gdsc.domain.membership.domain.MembershipValidator;
 import com.gdschongik.gdsc.domain.recruitment.dao.RecruitmentRoundRepository;
-import com.gdschongik.gdsc.domain.recruitment.domain.Recruitment;
 import com.gdschongik.gdsc.domain.recruitment.domain.RecruitmentRound;
 import com.gdschongik.gdsc.global.exception.CustomException;
 import com.gdschongik.gdsc.global.util.MemberUtil;
@@ -39,12 +38,16 @@ public class MembershipService {
     @Transactional
     public void submitMembership(Long recruitmentRoundId) {
         Member currentMember = memberUtil.getCurrentMember();
+
         RecruitmentRound recruitmentRound = recruitmentRoundRepository
                 .findById(recruitmentRoundId)
                 .orElseThrow(() -> new CustomException(RECRUITMENT_ROUND_NOT_FOUND));
 
-        validateMembershipDuplicate(currentMember, recruitmentRound.getRecruitment());
-        membershipValidator.validateMembershipSubmit(recruitmentRound);
+        Optional<Membership> existingMembership = membershipRepository.findAllByMember(currentMember).stream()
+                .filter(m -> m.getRecruitmentRound().getRecruitment().equals(recruitmentRound.getRecruitment()))
+                .findFirst();
+
+        membershipValidator.validateMembershipSubmit(recruitmentRound, existingMembership);
 
         Membership membership = Membership.createMembership(currentMember, recruitmentRound);
         membershipRepository.save(membership);
@@ -54,18 +57,5 @@ public class MembershipService {
 
     public Optional<Membership> findMyMembership(Member member, RecruitmentRound recruitmentRound) {
         return membershipRepository.findByMemberAndRecruitmentRound(member, recruitmentRound);
-    }
-
-    private void validateMembershipDuplicate(Member currentMember, Recruitment recruitment) {
-        Optional<Membership> membership = membershipRepository.findAllByMember(currentMember).stream()
-                .filter(m -> m.getRecruitmentRound().getRecruitment().equals(recruitment))
-                .findFirst();
-
-        if (membership.isPresent()) {
-            if (membership.get().isRegularRequirementAllSatisfied()) {
-                throw new CustomException(MEMBERSHIP_ALREADY_SATISFIED);
-            }
-            throw new CustomException(MEMBERSHIP_ALREADY_SUBMITTED);
-        }
     }
 }
